Improve the structure of tips class subclasses
Harmonize subclass structure. Add enum class for subclass types.
In addition, the tests were corrected to match the new structure.
