+++ src/main/java/remember/domain/InstanceType.java
@@ -1,0 +1,5 @@
+package remember.domain;
+
+public enum InstanceType {
+    BOOK, BLOGPOST, VIDEO
+}
