+++ src/main/java/remember/domain/instances/Blogpost.java
@@ -1,16 +1,19 @@
 package remember.domain.instances;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.NotBlank;
 import org.hibernate.validator.constraints.URL;
+import remember.domain.InstanceType;
 import remember.domain.Tip;
 
 import javax.persistence.Entity;
 
 @Entity
 @NoArgsConstructor
+@AllArgsConstructor
 @Data
 public class Blogpost extends Tip {
 
@@ -22,11 +25,11 @@ public class Blogpost extends Tip {
     @URL(message = "error.url.invalid")
     private String url;
 
-    public Blogpost(String author, String url, String title, String comment) {
-        this.author = author;
-        this.url = url;
+    public Blogpost(String title, String comment, String author, String url) {
         this.setTitle(title);
         this.setComment(comment);
-        this.setType("blogpost");
+        this.setAuthor(author);
+        this.setUrl(url);
+        this.setType(InstanceType.BLOGPOST);
     }
 }
