+++ src/test/java/remember/controller/BookRestControllerTest.java
@@ -69,12 +69,11 @@ public class BookRestControllerTest {
     public void setup() throws Exception {
         MockitoAnnotations.initMocks(this);
         this.mockMvc = MockMvcBuilders.webAppContextSetup(this.webApplicationContext).dispatchOptions(true).build();
-
         this.bookRepository.deleteAllInBatch();
 
-        this.book = bookRepository.save(new Book("Tom", "Tom's adventure", "fine book"));
-        this.bookList.add(bookRepository.save(new Book("J. K. Rowling", "Harry Potter and the Cursed Child", "A description")));
-        this.bookList.add(bookRepository.save(new Book("J. R. R. Tolkien", "The Lord of the Rings", "The Lord of the Rings is an epic high fantasy novel written by English author and scholar J. R. R. Tolkien.")));
+        this.book = bookRepository.save(new Book("Tom's adventure", "fine book", "Tom"));
+        this.bookList.add(bookRepository.save(new Book( "Harry Potter and the Cursed Child", "A description", "J. K. Rowling")));
+        this.bookList.add(bookRepository.save(new Book( "The Lord of the Rings", "The Lord of the Rings is an epic high fantasy novel written by English author and scholar J. R. R. Tolkien.", "J. R. R. Tolkien")));
     }
 
     @Test
